chore: debugger improvements 2 (#6494)
* feat: debugger builder and clean up terminal handling

* typos
